(refactor) updated allergies e2e tests to standards
