+++ src/applications/mhv-medical-records/tests/e2e/accelerated/fixtures/allergies/sample-lighthouse.json
