--- src/applications/mhv-medical-records/tests/e2e/accelerated/medical-records-view-all-domains-enabled.cypress.oracle-health.cypress.spec.js
@@ -1,76 +1,0 @@
-import MedicalRecordsSite from '../mr_site/MedicalRecordsSite';
-// import VitalsListPage from './pages/VitalsListPage';
-import oracleHealthUser from '../fixtures/user/oracle-health.json';
-
-describe('Medical Records View Vitals', () => {
-  const site = new MedicalRecordsSite();
-
-  beforeEach(() => {
-    site.login(oracleHealthUser, false);
-    site.mockFeatureToggles({
-      isAcceleratingEnabled: true,
-      isAcceleratingVitals: true,
-      isAcceleratingAllergies: true,
-    });
-    cy.intercept('POST', '/my_health/v1/medical_records/session', {}).as(
-      'session',
-    );
-    cy.intercept('GET', '/my_health/v1/medical_records/session/status', req => {
-      req.reply({
-        retrievedDate: 1732224967218,
-        lastRefreshDate: null,
-        facilityExtractStatusList: [
-          {
-            extract: 'Allergy',
-            lastRequested: 1732224367218,
-            lastCompleted: 1732224667218,
-            lastSuccessfulCompleted: 1732224667218,
-          },
-          {
-            extract: 'ImagingStudy',
-            lastRequested: 1732224367218,
-            lastCompleted: 1732224667218,
-            lastSuccessfulCompleted: 1732224667218,
-          },
-          {
-            extract: 'VPR',
-            lastRequested: 1732224367218,
-            lastCompleted: 1732224667218,
-            lastSuccessfulCompleted: 1732224667218,
-          },
-          {
-            extract: 'ChemistryHematology',
-            lastRequested: 1732224367218,
-            lastCompleted: 1732224667218,
-            lastSuccessfulCompleted: 1732224667218,
-          },
-        ],
-      });
-    });
-    cy.visit('my-health/medical-records');
-  });
-
-  it('Visits View Vitals List', () => {
-    // check for MY Va Health links
-    cy.get('[data-testid="labs-and-tests-oh-landing-page-link"]').should(
-      'be.visible',
-    );
-    cy.get('[data-testid="summary-and-notes-oh-landing-page-link"]').should(
-      'be.visible',
-    );
-    cy.get('[data-testid="vaccines-oh-landing-page-link"]').should(
-      'be.visible',
-    );
-    cy.get('[data-testid="health-conditions-oh-landing-page-link"]').should(
-      'be.visible',
-    );
-
-    cy.get('[data-testid="allergies-landing-page-link"]').should('be.visible');
-
-    cy.get('[data-testid="vitals-landing-page-link"]').should('be.visible');
-
-    // Axe check
-    cy.injectAxe();
-    cy.axeCheck('main');
-  });
-});
