+++ src/applications/mhv-medical-records/tests/e2e/accelerated/pages/Allergies.js
@@ -1,4 +1,52 @@
 class Allergies {
+  setIntercepts = ({ allergiesData, useOhData = true } = {}) => {
+    cy.intercept('POST', '/my_health/v1/medical_records/session', {}).as(
+      'session',
+    );
+    cy.intercept('GET', '/my_health/v1/medical_records/session/status', req => {
+      req.reply({
+        retrievedDate: 1732224967218,
+        lastRefreshDate: null,
+        facilityExtractStatusList: [
+          {
+            extract: 'Allergy',
+            lastRequested: 1732224367218,
+            lastCompleted: 1732224667218,
+            lastSuccessfulCompleted: 1732224667218,
+          },
+          {
+            extract: 'ImagingStudy',
+            lastRequested: 1732224367218,
+            lastCompleted: 1732224667218,
+            lastSuccessfulCompleted: 1732224667218,
+          },
+          {
+            extract: 'VPR',
+            lastRequested: 1732224367218,
+            lastCompleted: 1732224667218,
+            lastSuccessfulCompleted: 1732224667218,
+          },
+          {
+            extract: 'ChemistryHematology',
+            lastRequested: 1732224367218,
+            lastCompleted: 1732224667218,
+            lastSuccessfulCompleted: 1732224667218,
+          },
+        ],
+      });
+    });
+
+    cy.intercept('GET', '/my_health/v1/medical_records/allergies*', req => {
+      // check the correct param was used
+      if (useOhData) {
+        expect(req.url).to.contain('use_oh_data_path=1');
+      } else {
+        expect(req.url).to.not.contain('use_oh_data_path=1');
+      }
+      req.reply(allergiesData);
+    }).as('allergies-list');
+  };
+
   checkLandingPageLinks = () => {
     cy.get('[data-testid="labs-and-tests-oh-landing-page-link"]').should(
       'be.visible',
@@ -20,6 +68,12 @@ class Allergies {
     cy.get('[data-testid="allergies-landing-page-link"]')
       .should('be.visible')
       .click();
+  };
+
+  checkTitle = ({
+    title = 'Allergies and Reactions - Medical Records | Veterans Affairs',
+  }) => {
+    cy.title().should('contain', title);
   };
 }
